feat: pnpm store usages <package>...
* docs(pnpm): docs for the package usages feature

- Add documentation for `pnpm store usages`
- Refactor docs for `pnpm store` to be more readable

* feat(package-store): implement findPackageUsages

Introduces functionality to search the package store for packages.
Reports all matching packages and the node projects they are used in.
Allows for searching a specific version, or all versions.

* feat(server): expose findPackageUsages API

Exposes an API for finding package usages in the store.

* feat(store-controller-types): Exposes findPackageUsages store functionality

Publicly exposes the findPackageUsages function in the store-controller.
Also exposes the expected response type, allowing clients to be type-safe.

* feat(supi): implement findStoreUsages functionality, add API

Publicly exposes the findPackageUsages function SUPI.

* feat(pnpm): pnpm store usages

Enabled CLI command to find usages of a package in the global store.
Pretty-prints a tree with queries and results.

* style: remove semicolons

Removed semicolons to follow style guide.

* refactor(supi): change test after merge of #1528

* refactor: rename findPackageUsages return type

Better naming of the types returned in the pnpm store usages feature.

* tests: add tests for `pnpm store usages`

Add tests to `supi` to ensure datamodel returned by `package-store` is correct.
Fix test messages in `server`

PR #1529