feat: add support of the update-notifier configuration option (#4285)
The configuration option `update-notifier` allows users to disable the
update verification. This is interesting when pnpm is installed from
another package manager because the given instructions will not be
accurate. The `update-notifier` option exists in NPM so it can also
ease the migration to pnpm.
(https://docs.npmjs.com/cli/v8/using-npm/config#update-notifier).

close #4158.

Co-authored-by: Zoltan Kochan <z@kochan.io>