fix(cmd-shim): extend NODE_PATH with path to hidden hoisted dir (#4513)
Deprecating `extend-node-path` in pnpm v7 has caused issues with "next dev".

This change is bringing back extending `NODE_PATH` in bin command shims. However, only when `node-linker` is set to `isolated` and packages are hoisted to `node_modules/.pnpm/node_modules`. Only `node_modules/.pnpm/node_modules` is added to `NODE_PATH`, so it should not cause too long input errors (as it was sometimes the case in pnpm v6)