test(dependencies-hierarchy): fix height > requestedDepth cache test
Looking over this again, it looks like this never tested the fully
visited cache. The `a` package was never added to the fully visited
cache since the dependencies of its leaf nodes were never enumerated.

Adding a comment to make this more clear and increasing the max depth
of the test case.